refactor: router.dart 클린업
- sub-route path 하드코딩을 AppRoutes 상수로 추출
- ChangeNotifierProvider 설정을 router.dart로 일원화하고 LoginScreen 제거
- vjew_models 오타를 view_models로 수정
